api: fix errors in Player api introduced in commit 997f1bb4
Commit 997f1bb4 split the User table into two tables: User and Player.
The Player API needs to be updated to match this change.

The Player API is intended to allow game servers to verify that the token
provided by the player is valid.  When the API is called with the
user-supplied token, it returns the player's AGA ID and email address as
confirmation.  This method of calling the API is available for game server
and ratings server admins.

When the API is called with no arguments, it returns a list of all the
users registered in the AGA Ratings Server, along with their game server
account names.  This method of calling the API is only available to admins
on the ratings server.

Signed-off-by: Tim Henigan <tim.henigan@gmail.com>