Switch from using User.claimed to using User.fake to mark fake accounts
Using User.claimed was confusing because an AGA member may not
actually have any game records, or the AGA data may not have been
loaded yet. So the existence of a single User(aga_id=XXX, claimed=False)
entry in the db would not be able to distinguish between whether
it was a fake account or a real account.