Remove 8080 Verbage
Turns out that Flask Security has a bug that makes using a custom port
problematic. Seems like that particular library isn't being maintained
very well and has mostly been replaced by Flask-Security-Too. I tried
dropping that in, but the jump in version was too much to do simply,
so I just removed the wording suggesting that 8080 would be fine.

I think before this gets on production it make sense to upgrade most
of the libraries involved, but in the interest of getting some
momentum on the project this should be sufficient.

I had an errant piece of code that I accidentally checked in from the
last commit. It was from a failed approach to solving the problem, but
I missed it when I was cleaning up afterwards.